+++ HISTORY.md
@@ -1,5 +1,5 @@
-unreleased
-==========
+0.1.0 / 2014-06-17
+==================
 
   * Accept `req` as argument to `parse`
   * Accept `res` as argument to `parse`